+++ src/main/java/com/db/ayce/be/repository/UtenteRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.db.ayce.be.entity.Utente;
 
-public interface UtenteRepository extends JpaRepository<Utente, Integer> {
+public interface UtenteRepository extends JpaRepository<Utente, Long> {
     Optional<Utente> findByUsername(String username);
 }
