+++ src/main/java/com/db/ayce/be/service/ProdottoService.java
@@ -6,8 +6,8 @@ import com.db.ayce.be.entity.Prodotto;
 
 public interface ProdottoService {
     List<Prodotto> findAll();
-    Prodotto findById(Integer id);
+    Prodotto findById(Long id);
     Prodotto save(Prodotto prodotto);
-    Prodotto update(Integer id, Prodotto prodotto);
-    void delete(Integer id);
+    Prodotto update(Long id, Prodotto prodotto);
+    void delete(Long id);
 }
